+++ es_client/helpers/config.py
@@ -1,8 +1,8 @@
-"""Client builder helper functions"""
+"""Command-line configuration parsing and client builder helper functions"""
 from typing import Tuple
 import logging
 from shutil import get_terminal_size
-from click import secho
+from click import Context, secho
 from elasticsearch8 import Elasticsearch
 from es_client.builder import Builder, ClientArgs, OtherArgs
 from es_client import defaults
@@ -13,15 +13,105 @@ def cli_opts(
         value: str, settings: dict=None,
         onoff: dict=None, override: dict=None) -> Tuple[Tuple[str,], dict]:
     """
-    In order to make building a Click interface more cleanly, this function returns all Click
-    option settings indicated by ``value``, both forming the lone argument (e.g. ``--option``),
-    and all key word arguments as a dict.
+    :param value: The command-line :py:class:`option <click.Option>` name. The key must be present
+        in `settings`, or in :py:const:`CLICK_SETTINGS <es_client.defaults.CLICK_SETTINGS>`
+    :param settings: A dictionary consisting of :py:class:`click.Option` names as keys, with each
+        key having a dictionary consisting of :py:class:`click.Option` parameter names as keys, with
+        their associated settings as the value. If `settings` is not provided, it will be populated
+        by :py:const:`CLICK_SETTINGS <es_client.defaults.CLICK_SETTINGS>`.
+    :param onoff: A dictionary consisting of the keys `on` and `off`, with values used to set up a
+        `Click boolean option`_, .e.g. ``{'on': '', 'off': 'no-'}``. See below for examples.
+    :param override: A dictionary consisting of keys in `settings` with values you wish to override.
 
-    The single arg is rendered as ``f'--{value}'``. Likewise, ``value`` is the key to extract
-    all keyword args from the supplied dictionary.
-    The facilities to override default values and show hidden values is added here.
-    For default value overriding, the NOPE constant is used as None and False are valid default
-    values
+    :type value: str
+    :type settings: dict
+    :type onoff: dict
+    :type override: dict
+
+    :rtype: Tuple
+    :returns: A value suitable to use with the :py:func:`click.option` decorator, appearing as a
+        tuple containing a tuple and a dictionary, e.g. 
+
+        .. code-block:: python
+        
+          (('--OPTION1',),{'key1', 'value1', ...})
+
+    Click uses decorators to establish :py:class:`options <click.Option>` and
+    :py:class:`arguments <click.Argument>` for a :py:class:`command <click.Command>`. The parameters
+    specified for these decorator functions can be stored as default dictionaries, then expanded and
+    overridden, if desired.
+
+    In the `cli_example.py` file, the regular
+    :py:func:`click.option decorator function <click.option>` is wrapped by
+    :py:func:`option_wrapper() <es_client.helpers.utils.option_wrapper>`, and is aliased as
+    ``click_opt_wrap``. This wrapped decorator in turn calls this function and utilizes ``*``
+    arg expansion. If `settings` is `None`, default values from
+    :py:const:`CLICK_SETTINGS <es_client.defaults.CLICK_SETTINGS>`, are used to populate `settings`.
+    This function calls :func:`override_settings()` to override keys in `settings` with values from
+    matching keys in `override`.
+
+    In the example file, this looks like this:
+
+    .. code-block:: python
+
+      import click
+      from es_client.helpers.utils import option_wrapper
+      ONOFF = {'on': '', 'off': 'no-'}
+      click_opt_wrap = option_wrapper()
+
+      @click.group(context_settings=context_settings())
+      @click_opt_wrap(*cli_opts('OPTION1', settings={KEY: NEWVALUE}))
+      @click_opt_wrap(*cli_opts('OPTION2', onoff=ONOFF))
+      ...
+      @click_opt_wrap(*cli_opts('OPTIONX'))
+      @click.pass_context
+      def run(ctx, OPTION1, OPTION2, ..., OPTIONX):
+          # code here
+
+    The default setting KEY of ``OPTION1`` would be overriden by NEWVALUE.
+
+    ``OPTION2`` automatically becomes a `Click boolean option`_, which splits the option into an
+    enabled/disabled dichotomy by option name. In this example, it will be rendered as:
+    
+    .. code-block:: python
+
+      '--OPTION2/--no-OPTION2'
+
+    The dictionary structure of `ONOFF` is what this what this function requires, i.e. an `on` key
+    and an `off` key. The values for `on` and `off` can be whatever you like, e.g.  
+
+    .. code-block:: python
+
+      ONOFF = {'on': 'enable-', 'off': 'disable-'}
+
+    which would render as:
+    
+    .. code-block:: python
+    
+      '--enable-OPTION2/--disable-OPTION2'
+    
+    based on the above example.
+
+    It could also be:
+
+    .. code-block:: python
+
+      ONOFF = {'on': 'monty-', 'off': 'python-'}
+
+    which would render as: 
+    
+    .. code-block:: python
+    
+      '--monty-OPTION2/--python-OPTION2'
+    
+    but that would be too silly.
+
+    A :py:exc:`ConfigurationError <es_client.exceptions.ConfigurationError>` is raised `value` is
+    not found as a key in `settings`, or if the `onoff` parsing fails.
+    fails.
+
+    .. _Click boolean option: https://click.palletsprojects.com/en/8.1.x/options/#boolean-flags
+
     """
     if override is None:
         override = {}
@@ -39,93 +129,136 @@ def cli_opts(
             raise ConfigurationError from exc
     return (argval,), override_settings(settings[value], override)
 
-def cloud_id_override(args: dict, params: dict, client_args: ClientArgs) -> dict:
+def cloud_id_override(args: dict, ctx: Context) -> dict:
     """
-    If hosts are in the config file, but cloud_id is specified at the command-line,
-    we need to remove the hosts parameter as cloud_id and hosts are mutually exclusive
-
-    This function returns an updated dictionary ``args`` to be used for the final configuration as
-    well as updates the ``client_args`` object.
-
-    :param args: The argument dictionary derived params.
-    :param params: Click context params, e.g. ``ctx.params``
-    :param client_args: The working list of client args
+    :param args: A dictionary built from :py:attr:`ctx.params <click.Context.params>` keys and
+        values.
+    :param ctx: The Click command context 
 
     :type args: dict
-    :type params: dict from :py:class:`~.click.Context.params`
-    :type client_args: :py:class:`~.es_client.builder.ClientArgs`
+    :type ctx: :py:class:`Context <click.Context>`
 
     :rtype: dict
-    :returns: Updated version of ``args``
+    :returns: Updated version of `args`
+
+    If ``hosts`` are defined in the YAML configuration file, but ``cloud_id`` is specified at the
+    command-line, we need to remove the ``hosts`` parameter from the configuration dictionary built
+    from the YAML file before merging. Command-line provided arguments always supersede
+    configuration file ones. In this case, ``cloud_id`` and ``hosts`` are mutually exclusive, and
+    the command-line provided ``cloud_id`` must supersede a configuration file provided ``hosts``.
+    
+    This function returns an updated dictionary `args` to be used for the final configuration as
+    well as updates the :py:attr:`ctx.obj['client_args'] <click.Context.obj>` object. It's simply
+    easier to merge dictionaries using a separate object. It would be a pain and unnecessary to
+    make another entry in :py:attr:`ctx.obj <click.Context.obj>` for this.
     """
     logger = logging.getLogger(__name__)
-    if 'cloud_id' in params and params['cloud_id']:
-        logger.info('cloud_id from command-line superseding configuration file settings')
-        client_args.hosts = None
+    if 'cloud_id' in ctx.params and ctx.params['cloud_id']:
+        logger.debug('cloud_id from command-line superseding configuration file settings')
+        ctx.obj['client_args'].hosts = None
         args.pop('hosts', None)
     return args
 
-def context_settings():
-    """Return Click context settings dictionary"""
-    help_options = {'help_option_names': ['-h', '--help']}
-    return {**get_width(), **help_options}
-
-def get_arg_objects(config: dict) -> Tuple[ClientArgs, OtherArgs]:
+def context_settings() -> dict:
     """
-    Return initial tuple of :py:class:`~.es_client.builder.ClientArgs`, 
-    :py:class:`~.es_client.builder.OtherArgs`
-
-    They will be either empty/default, or with values from ``config``
-
-    :param config: Config dict derived from a YAML configuration file.
-
-    :type config: dict
+    :rtype: dict
+    :returns: A dictionary suitable to be used as the Click :py:class:`Command <click.Command>`
+        `context_settings` parameter.
     
-    :rtype: tuple
-    :returns: :py:class:`~.es_client.builder.ClientArgs` and 
-        :py:class:`~.es_client.builder.OtherArgs` objects in a tuple
+    Includes the terminal width from :py:func:`get_width()`
+    
+    Help format settings:
+    
+    .. code-block:: python
+    
+       {'help_option_names': ['-h', '--help']}
+    
+    And automatic environment variable reading based on a prefix value: 
+    
+    .. code-block:: python
+    
+       {'auto_envvar_prefix': ENV_VAR_PREFIX'}
+
+    from :py:const:`ENV_VAR_PREFIX <es_client.defaults.ENV_VAR_PREFIX>`
     """
-    client_args = ClientArgs()
-    other_args = OtherArgs()
-    validated_config = check_config(config)
-    client_args.update_settings(validated_config['client'])
-    other_args.update_settings(validated_config['other_settings'])
-    return client_args, other_args
+    prefix = {'auto_envvar_prefix': defaults.ENV_VAR_PREFIX}
+    help_options = {'help_option_names': ['-h', '--help']}
+    return {**get_width(), **help_options, **prefix}
 
-def get_args(params: dict, config: dict) -> Tuple[ClientArgs, OtherArgs]:
+def generate_configdict(ctx: Context) -> None:
     """
-    Return ClientArgs, OtherArgs tuple from params overriding whatever may be in config
+    :param ctx: The Click command context 
 
-    :param params: Click context params, e.g. ``ctx.params``
-    :param default_filepath: File path to a default config file location.
+    :type ctx: :py:class:`Context <click.Context>`
 
-    :type params: dict from :py:class:`~.click.Context.params``
-    :type default_filepath: str
+    :rtype: None
 
-    :rtype: tuple
-    :returns: :py:class:`~.es_client.builder.ClientArgs` and 
-        :py:class:`~.es_client.builder.OtherArgs` objects in a tuple
+    Generate a client configuration dictionary from :py:attr:`ctx.params <click.Context.params>` and
+    :py:attr:`ctx.obj['default_config'] <click.Context.obj>` (if provided), suitable for use as the
+    ``VALUE`` in :py:class:`Builder(configdict=VALUE) <es_client.builder.Builder>`
+    
+    It is stored as :py:attr:`ctx.obj['default_config'] <click.Context.obj>` and can be referenced
+    after this function returns.
+
+    The flow of this function is as follows:
+    
+    Step 1: Call :func:`get_arg_objects()` to create
+    :py:attr:`ctx.obj['client_args'] <click.Context.obj>` and
+    :py:attr:`ctx.obj['other_args'] <click.Context.obj>`, then update their values from
+    :py:attr:`ctx.obj['draftcfg'] <click.Context.obj>` (which was populated by
+    :func:`get_config()`).
+
+    Step 2: Call :func:`override_client_args()` and :func:`override_other_args()`, which
+    will use command-line args from :py:attr:`ctx.params <click.Context.params>` to override any
+    values from the YAML configuration file.
+
+    Step 3: Populate :py:attr:`ctx.obj['configdict'] <click.Context.obj>` from the resulting values.
     """
-    client_args, other_args = get_arg_objects(config)
-    override_client_args(params, client_args)
-    override_other_args(params, other_args)
-    return client_args, other_args
+    get_arg_objects(ctx)
+    override_client_args(ctx)
+    override_other_args(ctx)
+    ctx.obj['configdict'] = {
+        'elasticsearch': {
+            'client': prune_nones(ctx.obj['client_args'].asdict()),
+            'other_settings': prune_nones(ctx.obj['other_args'].asdict())
+        }
+    }
+
+def get_arg_objects(ctx: Context) -> None:
+    """
+    :param ctx: The Click command context 
+
+    :type ctx: :py:class:`Context <click.Context>`
+    
+    :rtype: None
+
+    Set :py:attr:`ctx.obj['client_args'] <click.Context.obj>` as a
+    :py:class:`~.es_client.builder.ClientArgs` object, and
+    :py:attr:`ctx.obj['other_args'] <click.Context.obj>` as an
+    :py:class:`~.es_client.builder.OtherArgs` object.
+
+    These will be updated with values returned from 
+    :func:`check_config(ctx.obj['draftcfg']) <es_client.helpers.utils.check_config>`.
+    
+    :py:attr:`ctx.obj['draftcfg'] <click.Context.obj>` was populated when :func:`get_config()` was
+    called.
+    """
+    ctx.obj['client_args'] = ClientArgs()
+    ctx.obj['other_args'] = OtherArgs()
+    validated_config = check_config(ctx.obj['draftcfg'], quiet=True)
+    ctx.obj['client_args'].update_settings(validated_config['client'])
+    ctx.obj['other_args'].update_settings(validated_config['other_settings'])
 
 def get_client(
     configdict: dict=None, configfile: str=None, autoconnect: bool=False,
     version_min: tuple=defaults.VERSION_MIN,
     version_max: tuple=defaults.VERSION_MAX) -> Elasticsearch:
-    """Get an Elasticsearch Client using :py:class:`es_client.builder.Builder`
-
-    Build a client out of settings from `configfile` or `configdict`
-    If neither `configfile` nor `configdict` is provided, empty defaults will be used.
-    If both are provided, `configdict` will be used, and `configfile` ignored.
-
+    """
     :param configdict: A configuration dictionary
-    :param configfile: A configuration file
+    :param configfile: A YAML configuration file
     :param autoconnect: Connect to client automatically
-    :param verion_min: Minimum acceptable version of Elasticsearch (major, minor, patch)
-    :param verion_max: Maximum acceptable version of Elasticsearch (major, minor, patch)
+    :param version_min: Minimum acceptable version of Elasticsearch (major, minor, patch)
+    :param version_max: Maximum acceptable version of Elasticsearch (major, minor, patch)
 
     :type configdict: dict
     :type configfile: str
@@ -135,6 +268,17 @@ def get_client(
 
     :returns: A client connection object
     :rtype: :py:class:`~.elasticsearch.Elasticsearch`
+
+    Get an Elasticsearch Client using :py:class:`~.es_client.builder.Builder`
+
+    Build a client connection object out of settings from `configfile` or `configdict`.
+
+    If neither `configfile` nor `configdict` is provided, empty defaults will be used.
+    
+    If both are provided, `configdict` will be used, and `configfile` ignored.
+
+    Raises :py:exc:`ESClientException <es_client.exceptions.ESClientException>` if unable to
+    connect.
     """
     logger = logging.getLogger(__name__)
     logger.debug('Creating client object and testing connection')
@@ -153,45 +297,59 @@ def get_client(
 
     return builder.client
 
-def get_config(params: dict, default_config: str = None) -> dict:
+def get_config(ctx: Context, quiet: bool=True) -> dict:
     """
-    If params['config'] is a valid path, return the validated dictionary from the YAML
+    :param ctx: The Click command context
+    :param quiet: If the default configuration file is being used, suppress the ``STDOUT`` message
+        indicating that.
 
-    If nothing has been provided to params['config'], but default_config is populated, use that.
+    :type ctx: :py:class:`Context <click.Context>`
+    :type quiet: bool
 
-    :param params: Click context params, e.g. ``ctx.params``
-    :param default_config: Path to a configuration file
+    :rtype: None
 
-    :type params: dict from :py:class:`~.click.Context.params``
-    :type default_config: str
+    If :py:attr:`ctx.params['config'] <click.Context.params>` is a valid path, return the validated
+    dictionary from the YAML.
 
-    :returns: Configuration dictionary
-    :rtype: dict
-    """
-    config = {'config':{}} # Set a default empty value
-    if params['config']:
-        config = get_yaml(params['config'])
-    # If no config was provided, but default config path exists, use it instead
-    elif default_config:
-        secho(f'Using default configuration file at {default_config}', bold=True)
-        config = get_yaml(default_config)
-    return config
-
-def get_hosts(params: dict) -> list:
-    """
-    Return hostlist suitable for client object. Validate url schema for each entry.
+    If nothing has been provided to :py:attr:`ctx.params['config'] <click.Context.params>`, but
+    :py:attr:`ctx.obj['default_config'] <click.Context.obj>` is populated, use that, and write a
+    line to ``STDOUT`` explaining this, unless `quiet` is `True`. 
     
-    :param params: Click context params, e.g. ``ctx.params``
+    Writing directly to ``STDOUT`` is done here because logging has not yet been configured, nor
+    can it be as the configuration options are just barely being read.
 
-    :type params: dict from :py:class:`~.click.Context.params``
+    Store the result in :py:attr:`ctx.obj['draftcfg'] <click.Context.obj>`
+    """
+    ctx.obj['draftcfg'] = {'config':{}} # Set a default empty value
+    if ctx.params['config']:
+        ctx.obj['draftcfg'] = get_yaml(ctx.params['config'])
+    # If no config was provided, but default config path exists, use it instead
+    elif 'default_config' in ctx.obj and ctx.obj['default_config']:
+        if not quiet:
+            secho(f"Using default configuration file at {ctx.obj['default_config']}", bold=True)
+        ctx.obj['draftcfg'] = get_yaml(ctx.obj['default_config'])
+    return ctx
 
-    :returns: List of hosts
+def get_hosts(ctx: Context) -> list:
+    """
+    :param ctx: The Click command context
+
+    :type ctx: :py:class:`Context <click.Context>`
+
+    :returns: A list of hosts
     :rtype: list
+
+    Return a list of hosts suitable for :py:attr:`ClientArgs.hosts <es_client.builder.ClientArgs>`
+    from :py:attr:`ctx.params['hosts'] <click.Context.params>`, validating the url schema for
+    Elasticsearch compliance for each host provided.
+
+    Raises a :py:exc:`ConfigurationError <es_client.exceptions.ConfigurationError>` if schema
+    validation fails.
     """
     logger = logging.getLogger(__name__)
     hostslist = []
-    if 'hosts' in params and params['hosts']:
-        for host in list(params['hosts']):
+    if 'hosts' in ctx.params and ctx.params['hosts']:
+        for host in list(ctx.params['hosts']):
             try:
                 hostslist.append(verify_url_schema(host))
             except ConfigurationError as err:
@@ -201,115 +359,165 @@ def get_hosts(params: dict) -> list:
         hostslist = None
     return hostslist
 
-def get_width():
-    """Determine terminal width"""
+def get_width() -> dict:
+    """
+    :rtype: dict
+    :returns: A dictionary suitable for use by itself as the Click
+        :py:class:`Command <click.Command>` `context_settings` parameter.
+
+    Determine terminal width by calling :py:func:`shutil.get_terminal_size`
+
+    Return value takes the form of ``{"max_content_width": get_terminal_size()[0]}``
+    """
     return {"max_content_width": get_terminal_size()[0]}
 
-def hosts_override(args: dict, params: dict, client_args: ClientArgs) -> dict:
+def hosts_override(args: dict, ctx: Context) -> dict:
     """
-    If hosts are provided at the command-line, but cloud_id was in the config file, we need to
-    remove the cloud_id parameter from the config file-based dictionary before merging.
-    This function returns an updated dictionary ``args`` to be used for the final configuration as
-    well as updates the ``client_args`` object.
-
-    :param args: The argument dictionary derived params.
-    :param params: Click context params, e.g. ``ctx.params``
-    :param client_args: The working list of client args
+    :param args: A dictionary built from :py:attr:`ctx.params <click.Context.params>` keys and
+        values.
+    :param ctx: The Click command context
 
     :type args: dict
-    :type params: dict from :py:class:`~.click.Context.params`
-    :type client_args: :py:class:`~.es_client.builder.ClientArgs`
+    :type ctx: :py:class:`Context <click.Context>`
 
     :rtype: dict
-    :returns: Updated version of ``args``
+    :returns: Updated version of `args`
+
+    If `hosts` are provided at the command-line and are present in
+    :py:attr:`ctx.params['hosts'] <click.Context.params>`, but `cloud_id` was in the config file,
+    we need to remove the `cloud_id` key from the configuration dictionary built from the YAML
+    file before merging. Command-line provided arguments always supersede configuration file ones,
+    including `hosts` overriding a file-based `cloud_id`.
+    
+    This function returns an updated dictionary `args` to be used for the final configuration as
+    well as updates the :py:attr:`ctx.obj['client_args'] <click.Context.obj>` object. It's simply
+    easier to merge dictionaries using a separate object. It would be a pain and unnecessary to
+    make another entry in :py:attr:`ctx.obj <click.Context.obj>` for this.
     """
     logger = logging.getLogger(__name__)
-    if 'hosts' in params and params['hosts']:
-        logger.info('hosts from command-line superseding configuration file settings')
-        client_args.hosts = None
-        client_args.cloud_id = None
+    if 'hosts' in ctx.params and ctx.params['hosts']:
+        logger.debug('hosts from command-line superseding configuration file settings')
+        ctx.obj['client_args'].hosts = None
+        ctx.obj['client_args'].cloud_id = None
         args.pop('cloud_id', None)
     return args
 
-def override_client_args(params: dict, client_args: ClientArgs) -> ClientArgs:
+def override_client_args(ctx: Context) -> None:
     """
-    Override client_args settings with any values found in params
-    
-    :param params: Click context params, e.g. ``ctx.params``
-    :param client_args: The working list of client args
+    :param ctx: The Click command context
 
-    :type params: dict from :py:class:`~.click.Context.params``
-    :type client_args: :py:class:`~.es_client.builder.ClientArgs`
+    :type ctx: :py:class:`Context <click.Context>`
 
-    :returns: The updated working list ClientArgs object
-    :rtype: :py:class:`~.es_client.builder.ClientArgs`
+    :rtype: None
+
+    Override :py:attr:`ctx.obj['client_args'] <click.Context.obj>` settings with any values found in
+        :py:attr:`ctx.params <click.Context.params>`
+
+    Update :py:attr:`ctx.obj['client_args'] <click.Context.obj>` with the results.
+
+    In the event that there are neither ``hosts`` nor a ``cloud_id`` after the updates, log to
+    debug that this is the case, and that the default value for ``hosts`` of
+    ``http://127.0.0.1:9200`` will be used.
     """
     logger = logging.getLogger(__name__)
     args = {}
-    # Populate args from params
-    for key, value in params.items():
-        if key in defaults.CLIENT_SETTINGS:
+    # Populate args from ctx.params
+    for key, value in ctx.params.items():
+        if key in defaults.config_settings():
             if key == 'hosts':
-                args[key] = get_hosts(params)
+                args[key] = get_hosts(ctx)
             elif value is not None:
                 args[key] = value
-    args = cloud_id_override(args, params, client_args)
-    args = hosts_override(args, params, client_args)
+    args = cloud_id_override(args, ctx)
+    args = hosts_override(args, ctx)
     args = prune_nones(args)
     # Update the object if we have settings to override after pruning None values
     if args:
-        client_args.update_settings(args)
+        ctx.obj['client_args'].update_settings(args)
     # Use a default hosts value of localhost:9200 if there is no host and no cloud_id set
-    if client_args.hosts is None and client_args.cloud_id is None:
+    if ctx.obj['client_args'].hosts is None and ctx.obj['client_args'].cloud_id is None:
         logger.debug('No hosts or cloud_id set! Setting default host to http://127.0.0.1:9200')
-        client_args.hosts = ["http://127.0.0.1:9200"]
-    return client_args
+        ctx.obj['client_args'].hosts = ["http://127.0.0.1:9200"]
 
-def override_other_args(params: dict, other_args: OtherArgs) -> OtherArgs:
+def override_other_args(ctx: Context) -> None:
     """
-    Override other_args settings with any values found in params
+    :param ctx: The Click command context
 
-    :param params: Click context params, e.g. ``ctx.params``
-    :param other_args: The working list of other args
+    :type ctx: :py:class:`Context <click.Context>`
 
-    :type params: dict from :py:class:`~.click.Context.params``
-    :type other_args: :py:class:`~.es_client.builder.OtherArgs`
+    :rtype: None
 
-    :returns: The updated working list OtherArgs object
-    :rtype: :py:class:`~.es_client.builder.OtherArgs`
+    Override :py:attr:`ctx.obj['other_args'] <click.Context.obj>` settings with any values found in
+    :py:attr:`ctx.params <click.Context.params>`
+
+    Update :py:attr:`ctx.obj['other_args'] <click.Context.obj>` with the results.
     """
-    args = prune_nones({
-        'master_only': params['master_only'],
-        'skip_version_test': params['skip_version_test'],
-        'username': params['username'],
-        'password': params['password'],
-        'api_key': {
-            'id': params['id'],
-            'api_key': params['api_key'],
-            'token': params['api_token'],
-        }
+    apikey = prune_nones({
+        'id': ctx.params['id'],
+        'api_key': ctx.params['api_key'],
+        'token': ctx.params['api_token'],
     })
+    args = prune_nones({
+        'master_only': ctx.params['master_only'],
+        'skip_version_test': ctx.params['skip_version_test'],
+        'username': ctx.params['username'],
+        'password': ctx.params['password'],
+    })
+    args['api_key'] = apikey
 
     # Remove `api_key` root key if `id` and `api_key` and `token` are all None
-    if params['id'] is None and params['api_key'] is None and params['api_token'] is None:
+    if (ctx.params['id'] is None and ctx.params['api_key'] is None and
+            ctx.params['api_token'] is None):
         del args['api_key']
 
     if args:
-        other_args.update_settings(args)
-    return other_args
+        ctx.obj['other_args'].update_settings(args)
 
 def override_settings(settings: dict, override: dict) -> dict:
     """
-    Override keys in settings with values matching in override
-
     :param settings: The source data
-    :param override: The data which will override ``settings``
+    :param override: The data which will override `settings`
 
     :type settings: dict
     :type override: dict
 
     :rtype: dict
-    :returns: An dictionary based on ``settings`` updated with values from ``override``
+    :returns: An dictionary based on `settings` updated with values from `override`
+
+    This function is called by :func:`cli_opts()` in order to override settings used in a
+    :py:class:`Click Option <click.Option>`.
+
+    Click uses decorators to establish :py:class:`options <click.Option>` and
+    :py:class:`arguments <click.Argument>` for a :py:class:`command <click.Command>`. The parameters
+    specified for these decorator functions can be stored as default dictionaries, then expanded and
+    overridden, if desired.
+
+    In the `cli_example.py` file, the regular
+    :py:func:`click.option decorator function <click.option>` is wrapped by
+    :py:func:`option_wrapper() <es_client.helpers.utils.option_wrapper>`, and is aliased as
+    ``click_opt_wrap``. This wrapped decorator in turn calls :func:`cli_opts()` and utilizes ``*``
+    arg expansion. :func:`cli_opts()` references defaults, and calls this function to override 
+    keys in `settings` with values from matching keys in `override`.
+
+    In the example file, this looks like this:
+
+    .. code-block:: python
+
+      import click
+      from es_client.helpers.utils import option_wrapper
+      OVERRIDE = {KEY: NEWVALUE}
+      click_opt_wrap = option_wrapper()
+
+      @click.group(context_settings=context_settings())
+      @click_opt_wrap(*cli_opts('OPTION1'))
+      @click_opt_wrap(*cli_opts('OPTION2', settings=OVERRIDE))
+      ...
+      @click_opt_wrap(*cli_opts('OPTIONX'))
+      @click.pass_context
+      def run(ctx, OPTION1, OPTION2, ..., OPTIONX):
+          # code here
+
+    The default setting KEY of ``OPTION2`` would be overriden by NEWVALUE.
     """
     if not isinstance(override, dict):
         raise ConfigurationError(f'override must be of type dict: {type(override)}')